# [Xamarin.Android.Build.Tasks] <GenerateJavaStubs/> improvements