I went through the `_GenerateJavaStubs` MSBuild target and the
`<GenerateJavaStubs/>` MSBuild task. There were a few more "easy wins"
to improve performance.

## Inputs & Outputs ##

`_GenerateJavaStubs` has `Inputs` of:

    Inputs="$(MSBuildAllProjects);@(_ResolvedAssemblies);$(_AndroidManifestAbs);$(_AndroidBuildPropertiesCache);@(_AndroidResourceDest)"

Lets say a NetStandard assembly changes, such as a XAML change in a
Xamarin.Forms app. In this case `_GenerateJavaStubs` does not actually
need to run again.

We can use `@(_ResolvedUserMonoAndroidAssemblies)` as an input instead
of `@(_ResolvedAssemblies)`.

This is the biggest win--to skip this target completely.

## Mono.Cecil loading on-demand ##

Previously we were calling `DirectoryAssemblyResolver.Load()` up front
on every assembly in `ResolvedAssemblies`. Instead, we could just add
their paths to `SearchDirectories` and let Mono.Cecil load them on
demand.

`JavaTypeScanner` is passed a subset of the assemblies it looks at,
which are `@(_ResolvedUserMonoAndroidAssemblies)` and a few
special-cased ones. In some cases not all assemblies will get loaded,
so we save a bit of time here.

## selectedWhitelistAssemblies ##

`ManifestDocument` had the concept of a "whitelist" assembly, which
currently consisted of a single assembly:

    internal static readonly string [] FrameworkAttributeLookupTargets = {"Mono.Android.GoogleMaps.dll"};

`Mono.Android.GoogleMaps.dll` no longer exists, so I was able to
remove a bit of code that still used that dead codepath.

## Typemaps ##

Just a couple minor improvements here:

* It looks like to me we can just use the constructor of
  `TypeNameMapGenerator` where we pass in the list of
  `TypeDefinition`.
* Fixed the usage of the `[Obsolete]` constructor. Using
  `this.CreateTaskLogger()` instead.
* I reused the `MemoryStream` and got rid of the `UpdateWhenChanged`
  method.

## Results ##

Testing the Xamarin.Forms project in this repo.

A fresh build:

    Before:
      1005 ms  _GenerateJavaStubs                         1 calls
    After:
       498 ms  _GenerateJavaStubs                         1 calls

An incremental build with XAML-change:

    Before:
       875 ms  _GenerateJavaStubs                         1 calls
    After:
      n/a

The target is skipped now:

    _GenerateJavaStubs:
    Skipping target "_GenerateJavaStubs" because all output files are up-to-date with respect to the input files.

Overall this seems to be ~500ms better for first builds, and ~800ms
better for incremental builds with XAML changes.