@jonpryor 2019-04-11
It doesn't immediately make sense, but this change seems to have broken ~every on-device `.apk` test: https://jenkins.mono-project.com/job/xamarin-android-pr-pipeline-release/518/testReport/ e.g. https://jenkins.mono-project.com/job/xamarin-android-pr-pipeline-release/518/testReport/Xamarin.Android/Locale_Tests/Possible_Crash___Release/

```
Unhandled Exception:
System.ArgumentException: Handle must be valid.
Parameter name: instance
  at Java.Interop.JniEnvironment+InstanceMethods.CallObjectMethod (Java.Interop.JniObjectReference instance, Java.Interop.JniMethodInfo method)
  at Android.Runtime.JNIEnv.CallObjectMethod (System.IntPtr jobject, System.IntPtr jmethod)
  at Java.Interop.TypeManager.GetClassName (System.IntPtr class_ptr)
  at Java.Interop.TypeManager.CreateInstance (System.IntPtr handle, Android.Runtime.JniHandleOwnership transfer, System.Type targetType)
  at Java.Lang.Object.GetObject (System.IntPtr handle, Android.Runtime.JniHandleOwnership transfer, System.Type type)
  at Java.Lang.Object._GetObject[T] (System.IntPtr handle, Android.Runtime.JniHandleOwnership transfer)
  at Java.Lang.Object.GetObject[T] (System.IntPtr handle, Android.Runtime.JniHandleOwnership transfer)
  at Java.Lang.Thread.get_DefaultUncaughtExceptionHandler ()
  at Android.Runtime.JNIEnv.Initialize (Android.Runtime.JnienvInitializeArgs* args)
```

Looks like the `typemap.*` files are missing a type?