@jonathanpeppers 2019-04-16
I _think_ the macOS failures are OK:
```
Test Result (2 failures / +1)
JnienvTest.Mono.Android_Tests, Java.InteropTests.JnienvTest.DoNotLeakWeakReferences / 
ReleaseXamarin.Android.Bcl_Tests.nunit.Possible Crash / Release
```
It seems like the "adb crashed" thing happened.